Fix printf format issues. (#379)
These issues appear when -Werror=format on 32-bit platforms depending on
the exact definitions of uint64_t and size_t.

It doesn't appear that bochs uses %zu for size_t -- since you support
Windows platforms, unclear if that would work for you. If so, that's
definitely correct rather than the unsigned long cast.