widen 'bytes' column (#411)
The 'bytes' column is automatically resized to the width of the word
'bytes' (its header). This makes most instruction rows show `...`.

This patch defaults the width to 125 pixels allowing you to resize it
during each session. However, every new session will default back to
125. A new entry in the `bx_enh_dbg.ini` file could be created to save
and restore this size instead of defaulting to 125 each time.

This patch is for Windows only. Other platforms may require other
modifications.