Vmware3 (#476)
If the vmware3 class doesn't find a valid vmware3 image file, it will
exit without removing the .lock file. A simple addition of
```
  bx_close_image(file_descriptor, pathname);
```
at line 257 would fix it:

https://github.com/bochs-emu/Bochs/blob/2663ff698c83f067035c6ab309a24561efff73d4/bochs/iodev/hdimage/vmware3.cc#L255-L259

However, to be more consistent with vmware4 I have tried to add a more
consistent solution. I did not check to see if this works with a valid
vmware3 file. I have none to test with. Please test...